+++ src/main/java/bussiness/payments/MobilePayment.java
@@ -6,10 +6,10 @@ import java.math.BigDecimal;
 
 public class MobilePayment extends Payment {
 
-    public MobilePayment(String mobileNumber, double amount) {
+    public MobilePayment(String mobileNumber, BigDecimal amount) {
 
         super();
         super.setMobileNumber(mobileNumber);
-        super.setAmount(BigDecimal.valueOf(amount));
+        super.setAmount(amount);
     }
 }
